feat(agent_egress_service): serve the egress proxy from its own binary
The proxy has had its own listener since it was written - one binary
today, as the comment in the harness's main put it. This makes that
true: the same `agent_egress` router, unchanged, behind a binary that
takes only what a credential-stamping proxy needs (Postgres, the
Pipedream rows, the GitHub App and Macro API signing keys) and none of
the harness's Kafka, Redis, Daytona or Cursor wiring.

The harness keeps `EgressProvisioner` - minting tokens and building
advertised URLs is still its job. Only the serving side leaves, taking
the secrets that existed solely to mint tokens inline with it.
