+++ .github/services-config.json
@@ -14,6 +14,19 @@
         "agent_harness_service"
       ]
     },
+    "agent-egress-service": {
+      "source_paths": [
+        "services/agent_egress_service/**",
+        "crates/agent_egress/**",
+        "crates/agent_session/**",
+        "crates/agent_fold/**",
+        "crates/agent_runtime_protocol/**"
+      ],
+      "stack_path": "infra/stacks/agent-egress-service/**",
+      "deploy_binaries": [
+        "agent_egress_service"
+      ]
+    },
     "agent-schedule-service": {
       "source_paths": [
         "services/scheduled_action/**"
