+++ .github/workspace-dep-closures.json
@@ -129,6 +129,98 @@
       "crates/user_quota",
       "crates/workspace-hack"
     ],
+    "agent_egress_service": [
+      "crates/activity",
+      "crates/agent",
+      "crates/agent_egress",
+      "crates/agent_fold",
+      "crates/agent_runtime_protocol",
+      "crates/agent_session",
+      "crates/ai_toolset",
+      "crates/ai_usage",
+      "crates/attachment",
+      "crates/bot_id",
+      "crates/bot_token",
+      "crates/bots",
+      "crates/broadcast",
+      "crates/channel_sender",
+      "crates/channels",
+      "crates/connection",
+      "crates/connection_gateway_client",
+      "crates/connection_gateway_models",
+      "crates/contacts",
+      "crates/cowlike",
+      "crates/database_env_vars",
+      "crates/document_sub_type",
+      "crates/documents",
+      "crates/embedding",
+      "crates/entity_access",
+      "crates/entity_access_db_utils",
+      "crates/entity_access_management",
+      "crates/entity_mutation",
+      "crates/filter_ast",
+      "crates/foreign_entity",
+      "crates/frecency",
+      "crates/fusionauth",
+      "crates/github",
+      "crates/harness_id",
+      "crates/harness_token",
+      "crates/invite_email",
+      "crates/item_filters",
+      "crates/kafka_util",
+      "crates/last_online_tracker",
+      "crates/lexical_client",
+      "crates/macro_auth",
+      "crates/macro_authorization",
+      "crates/macro_aws_config",
+      "crates/macro_config",
+      "crates/macro_config_derive",
+      "crates/macro_db_client",
+      "crates/macro_entrypoint",
+      "crates/macro_env",
+      "crates/macro_env_var",
+      "crates/macro_event_broker",
+      "crates/macro_event_topics",
+      "crates/macro_service_urls",
+      "crates/macro_sync_service_jwt",
+      "crates/macro_tower_layers",
+      "crates/macro_user_id",
+      "crates/macro_uuid",
+      "crates/mcp_toolset",
+      "crates/mention_utils",
+      "crates/model",
+      "crates/model-entity",
+      "crates/model-error-response",
+      "crates/model_file_type",
+      "crates/model_notifications",
+      "crates/model_user",
+      "crates/models_bulk_upload",
+      "crates/models_dcs",
+      "crates/models_entity_access_management",
+      "crates/models_opensearch",
+      "crates/models_pagination",
+      "crates/models_permissions",
+      "crates/models_properties",
+      "crates/models_team",
+      "crates/non_empty",
+      "crates/notification",
+      "crates/pipedream_mcp",
+      "crates/prompt",
+      "crates/rate_limit",
+      "crates/remote_env_var",
+      "crates/s3_key",
+      "crates/secretsmanager_client",
+      "crates/share_permission_db_utils",
+      "crates/sqs_worker",
+      "crates/sync_service",
+      "crates/sync_service_client",
+      "crates/system_properties",
+      "crates/system_skills",
+      "crates/task_dedup",
+      "crates/user_quota",
+      "crates/workspace-hack",
+      "services/agent_egress_service"
+    ],
     "agent_fold": [
       "crates/agent_fold",
       "crates/agent_runtime_protocol",
@@ -370,7 +462,6 @@
     "agent_inmem": [
       "crates/activity",
       "crates/agent",
-      "crates/agent_egress",
       "crates/agent_fold",
       "crates/agent_inmem",
       "crates/agent_runtime_protocol",
@@ -412,9 +503,7 @@
       "crates/filter_ast",
       "crates/foreign_entity",
       "crates/frecency",
-      "crates/fusionauth",
       "crates/generic_email_domains",
-      "crates/github",
       "crates/harness_id",
       "crates/harness_token",
       "crates/import",
