+++ services/agent_harness_service/src/main.rs
@@ -17,17 +17,8 @@ mod model_providers;
 mod runtime_commands;
 mod trigger;
 
-#[cfg(test)]
-mod test;
-
 use std::{future::Future, pin::Pin, sync::Arc};
 
-use agent_egress::domain::service::EgressServiceImpl;
-use agent_egress::outbound::forwarder::ReqwestForwarder;
-use agent_egress::outbound::github_tokens::GithubAppTokens;
-use agent_egress::outbound::macro_mcp::{MacroApiTokenSigner, WithMacroMcp};
-use agent_egress::outbound::mcp_credentials::PipedreamMcpCredentials;
-use agent_egress::outbound::session_authority::StoredTokenSessionAuthority;
 use agent_fold::domain::service::FoldedMessageService;
 use agent_harness::domain::model::{
     AgentKind, AgentRuntimeConfig, HarnessCommand, HarnessDefaults, SessionDefaults,
@@ -56,7 +47,6 @@ use agent_harness::outbound::routing::RoutedContainerManager;
 use agent_harness::outbound::runtime_registry::{HarnessKeyedConnections, RuntimeRegistry};
 use agent_inmem::domain::engine::TurnEngine;
 use agent_inmem::outbound::acp_mcp::AcpMcpConnector;
-use agent_inmem::outbound::egress_mcp::EgressMcpClient;
 use agent_inmem::outbound::log_frames::LogFrameSource;
 use agent_inmem::outbound::manager::InMemAgentManager;
 use agent_inmem::rig_engine::RigTurnEngine;
@@ -88,9 +78,6 @@ use containers::{InMemRuntime, RoutedContainers};
 use cursor_api_key::cipher::{AwsKmsCiphertexts, KmsCursorApiKeyCipher};
 use cursor_cloud_agents::api::CURSOR_API_BASE_URL;
 use cursor_cloud_agents::domain::model::RepoUrl as CursorRepoUrl;
-use github::domain::service::{InstallationTokenConfig, InstallationTokenService};
-use github::outbound::github_sync_client::GithubSyncClientImpl;
-use github::outbound::pg_github_sync_repo::PgGithubSyncRepo;
 use harness_bindings::{PgHarnessBindings, PgHarnessPresence};
 use harnesses::outbound::pg_harness_repo::PgHarnessRepo;
 use kafka_util::{GroupName, KafkaEventConsumer, consumer_span, record_span_error};
@@ -106,11 +93,8 @@ use macro_event_broker::{
     KafkaConsumerAdapter, KafkaEventPublisher, MacroEvent as _, MacroEventBrokerService,
     MacroEventCollection as _, MacroEventConsumerService,
 };
-use macro_service_urls::{
-    AgentHarnessEgressUrl, ConnectionGatewayUrl, LexicalServiceUrl, McpServiceUrl,
-};
+use macro_service_urls::{AgentHarnessEgressUrl, ConnectionGatewayUrl, LexicalServiceUrl};
 use model_providers::{CursorModels, InMemoryModels, MacrodModels, VisibleHarnessAccess};
-use pipedream_mcp::outbound::api::{PipedreamClient, PipedreamConfig};
 use pipedream_mcp::outbound::pg_connection_repo::PgConnectionRepo;
 use rdkafka::consumer::CommitMode;
 use rdkafka::message::{BorrowedMessage, Message as _};
@@ -163,22 +147,15 @@ async fn main() -> anyhow::Result<()> {
     result
 }
 
-fn macro_mcp_endpoint(base_url: &McpServiceUrl) -> Result<url::Url, url::ParseError> {
-    // Append rather than Url::join("/mcp"), which would discard the gateway prefix.
-    url::Url::parse(&format!("{}/mcp", base_url.trim_end_matches('/')))
-}
-
 async fn run() -> anyhow::Result<()> {
     agent_harness::install_tls_provider();
-    // AWS first, because the config's secrets resolve through Secrets Manager.
+    // AWS first: the JWT validator reads its public key from Secrets Manager
+    // and the Cursor key cipher decrypts through KMS.
     let aws_config = macro_aws_config::get_macro_aws_config().await;
     let secrets = secretsmanager_client::SecretsManager::new(aws_sdk_secretsmanager::Client::new(
         &aws_config,
     ));
-    let config = Config::from_env()?
-        .resolve_remote_secrets(Environment::new_or_prod(), &secrets)
-        .await
-        .context("failed to resolve agent harness service secrets")?;
+    let config = Config::from_env()?;
     let bot_id = BotId::new_from_uuid(config.harness_bot_id);
     let enable_dev_commands = matches!(
         config.environment,
@@ -289,59 +266,9 @@ async fn run() -> anyhow::Result<()> {
     let event_broker_tracker = tokio_util::task::TaskTracker::new();
     // MCP connections: the same rows the chat tool path reads, so an app
     // connected in Macro is an app the sandbox can reach, with nothing to
-    // keep in sync. The rows hold no secrets - Pipedream owns the grants.
+    // keep in sync. Read only to name the servers a session may dial; the
+    // egress service is what spends them.
     let mcp_connections = Arc::new(PgConnectionRepo::new(pool.clone()));
-
-    // The client that addresses Pipedream's remote MCP server, built from the
-    // same credentials `document_cognition_service` uses.
-    let pipedream = PipedreamClient::new(PipedreamConfig {
-        client_id: config.pipedream_client_id.to_string(),
-        client_secret: config.pipedream_client_secret.to_string(),
-        project_id: config.pipedream_project_id.to_string(),
-        environment: config.pipedream_environment.clone(),
-        api_url: config.pipedream_api_url.clone(),
-        mcp_url: config.pipedream_mcp_url.clone(),
-        // Only Connect tokens carry allowed origins, and this service never
-        // mints one: connecting apps stays in the app.
-        allowed_origins: Vec::new(),
-    })
-    .context("failed to build Pipedream client")?;
-
-    // Every session's MCP servers: Macro's own under the reserved `macro`
-    // slug, then the owner's Pipedream connections. The `macro` credential is
-    // signed inline with the same key authentication_service holds; what this
-    // process hands out is always single-user and minutes from expiry.
-    let mcp_credentials = WithMacroMcp::new(
-        PipedreamMcpCredentials::new(Arc::clone(&mcp_connections), pipedream),
-        MacroApiTokenSigner::new(
-            pool.clone(),
-            config.macro_api_token_issuer.as_ref(),
-            config.macro_api_token_private_secret_key.as_ref(),
-        ),
-        macro_mcp_endpoint(&McpServiceUrl::new()?).context("MCP service endpoint is not a URL")?,
-        // The one gate on cleartext: a local stack's mcp-service is dialed
-        // across the compose bridge, where TLS would be theater. Everywhere
-        // else, an http URL refuses to boot.
-        matches!(config.environment, Environment::Local),
-    )
-    .context("the macro MCP upstream is misconfigured")?;
-
-    // The egress proxy: one binary today, its own listener from the start.
-    // Shared with the in-memory runtime, which calls it directly rather than
-    // through that listener.
-    let egress = Arc::new(EgressServiceImpl::new(
-        StoredTokenSessionAuthority::new(PgAgentSessionRepo::new(pool.clone())),
-        mcp_credentials,
-        GithubAppTokens::new(InstallationTokenService::new(
-            InstallationTokenConfig {
-                client_id: config.github_sync_app_client_id.clone(),
-                private_key_pem: config.github_sync_app_pem_secret_key.as_ref().to_owned(),
-            },
-            PgGithubSyncRepo::new(pool.clone()),
-            GithubSyncClientImpl::default(),
-        )),
-        ReqwestForwarder::new()?,
-    ));
 
     let mut inmem_model_engine: Option<Arc<dyn TurnEngine>> = None;
     let inmem = match inmem_bot {
@@ -361,9 +288,7 @@ async fn run() -> anyhow::Result<()> {
                 manager: InMemAgentManager::new(
                     engine,
                     frames,
-                    Arc::new(AcpMcpConnector::new(EgressMcpClient::new(Arc::clone(
-                        &egress,
-                    )))),
+                    Arc::new(AcpMcpConnector::default()),
                 )
                 .with_dev_commands(enable_dev_commands),
             })
@@ -688,13 +613,6 @@ async fn run() -> anyhow::Result<()> {
         config.internal_api_key.clone(),
     ));
 
-    let egress_port = config.egress_port;
-    let egress_http = tokio::spawn(async move {
-        if let Err(error) = api::serve_egress(egress, egress_port, shutdown_signal()).await {
-            tracing::error!(error = ?error, "agent harness service egress stopped");
-        }
-    });
-
     // The consumer: every agent-session event, filtered to our bot.
     let consumer =
         KafkaEventConsumer::<AgentHarnessConsumerGroup>::from_env(config.kafka_brokers.as_ref())?;
@@ -873,7 +791,6 @@ async fn run() -> anyhow::Result<()> {
 
     http.abort();
     trigger.abort();
-    egress_http.abort();
     heartbeat.abort();
     runtime_commands.abort();
     let stop_failures = container_shutdown.shutdown_all().await;
